Allow custom space char width per hu_textline_t
The DIG* character graphics are 5 pixels wide, but previously spaces in
a textline were always drawn 4 pixels wide. This made it difficult to
create HUD text overlays in monospace, which is useful for padding and
aligning items consistently.

With this change, each textline has a variable called spcwidth that
defaults to 4 and specifies the width that space characters (i.e. ASCII
0x20) on that line should be drawn as, in pixels.

There are a lot of hardcoded '4' literals but I think I got all the ones
that represented the hardcoded space width and replaced them with a
reference to the textline's spcwidth. I might have missed a few
(hu_lib.c:581 maybe?) but it won't cause misbehavior on any widget that
uses the default value.